+++ demo/api/src/config/environment-variables.ts
@@ -1,7 +1,7 @@
-import { BlobStorageConfig } from "@comet/cms-api";
+import { BlobStorageConfig, IsUndefinable } from "@comet/cms-api";
 import { PrivateSiteConfig } from "@src/site-configs";
 import { Transform, Type } from "class-transformer";
-import { IsArray, IsBoolean, IsInt, IsOptional, IsString, IsUrl, MinLength, ValidateIf } from "class-validator";
+import { IsArray, IsBoolean, IsEmail, IsInt, IsOptional, IsString, IsUrl, MinLength, ValidateIf } from "class-validator";
 
 export class EnvironmentVariables {
     @IsString()
@@ -77,9 +77,6 @@ export class EnvironmentVariables {
     @IsString()
     IMGPROXY_KEY: string;
 
-    @IsInt()
-    IMGPROXY_QUALITY = 80;
-
     @IsString()
     @MinLength(16)
     DAM_SECRET: string;
@@ -121,6 +118,18 @@ export class EnvironmentVariables {
     @ValidateIf((v) => v.BLOB_STORAGE_DRIVER === "s3")
     @IsString()
     S3_BUCKET: string;
+
+    @IsUndefinable()
+    @IsArray()
+    @Transform(({ value }) => value.split(","))
+    @IsEmail({}, { each: true })
+    MAILER_SEND_ALL_MAILS_TO?: string[];
+
+    @IsUndefinable()
+    @IsArray()
+    @Transform(({ value }) => value.split(","))
+    @IsEmail({}, { each: true })
+    MAILER_SEND_ALL_MAILS_BCC?: string[];
 
     @IsString()
     @ValidateIf(() => process.env.NODE_ENV === "production")