+++ packages/api/api-generator/src/commands/generate/generateCrud/__tests__/generate-crud-dedicated-resolver-arg.spec.ts
@@ -3,11 +3,11 @@ import { BaseEntity, Collection, defineConfig, Entity, ManyToOne, MikroORM, OneT
 import { LazyMetadataStorage } from "@nestjs/graphql/dist/schema-builder/storages/lazy-metadata.storage";
 import { v4 as uuid } from "uuid";
 
-import { formatGeneratedFiles, parseSource } from "../../utils/test-helper";
+import { formatGeneratedFiles, parseSource, testPermission } from "../../utils/test-helper";
 import { generateCrud } from "../generate-crud";
 
 @Entity()
-@CrudGenerator({ targetDirectory: __dirname })
+@CrudGenerator({ targetDirectory: __dirname, requiredPermission: testPermission })
 class TestEntityProductVariant extends BaseEntity {
     @PrimaryKey({ columnType: "text", type: "string" })
     id: string;
@@ -21,7 +21,7 @@ class TestEntityProductVariant extends BaseEntity {
 }
 
 @Entity()
-@CrudGenerator({ targetDirectory: __dirname })
+@CrudGenerator({ targetDirectory: __dirname, requiredPermission: testPermission })
 class TestEntityProduct extends BaseEntity {
     @PrimaryKey({ type: "uuid" })
     id: string = uuid();
@@ -45,7 +45,10 @@ describe("GenerateCrud dedicatedResolverArg", () => {
                 }),
             );
 
-            const out = await generateCrud({ targetDirectory: __dirname }, orm.em.getMetadata().get("TestEntityProductVariant"));
+            const out = await generateCrud(
+                { targetDirectory: __dirname, requiredPermission: testPermission },
+                orm.em.getMetadata().get("TestEntityProductVariant"),
+            );
             const formattedOut = await formatGeneratedFiles(out);
             const file = formattedOut.find((file) => file.name === "dto/test-entity-product-variant.input.ts");
             if (!file) throw new Error("File not found");
@@ -71,7 +74,10 @@ describe("GenerateCrud dedicatedResolverArg", () => {
                 }),
             );
 
-            const out = await generateCrud({ targetDirectory: __dirname }, orm.em.getMetadata().get("TestEntityProductVariant"));
+            const out = await generateCrud(
+                { targetDirectory: __dirname, requiredPermission: testPermission },
+                orm.em.getMetadata().get("TestEntityProductVariant"),
+            );
             const formattedOut = await formatGeneratedFiles(out);
             const file = formattedOut.find((file) => file.name === "dto/test-entity-product-variants.args.ts");
             if (!file) throw new Error("File not found");
@@ -98,7 +104,10 @@ describe("GenerateCrud dedicatedResolverArg", () => {
                 }),
             );
 
-            const out = await generateCrud({ targetDirectory: __dirname }, orm.em.getMetadata().get("TestEntityProductVariant"));
+            const out = await generateCrud(
+                { targetDirectory: __dirname, requiredPermission: testPermission },
+                orm.em.getMetadata().get("TestEntityProductVariant"),
+            );
             const formattedOut = await formatGeneratedFiles(out);
             const file = formattedOut.find((file) => file.name === "test-entity-product-variant.resolver.ts");
             if (!file) throw new Error("File not found");