+++ packages/admin/admin-color-picker/CHANGELOG.md
@@ -117,6 +117,27 @@
     - @comet/admin@8.0.0-beta.0
     - @comet/admin-icons@8.0.0-beta.0
 
+## 7.25.3
+
+### Patch Changes
+
+- @comet/admin@7.25.3
+- @comet/admin-icons@7.25.3
+
+## 7.25.2
+
+### Patch Changes
+
+- @comet/admin@7.25.2
+- @comet/admin-icons@7.25.2
+
+## 7.25.1
+
+### Patch Changes
+
+- @comet/admin@7.25.1
+- @comet/admin-icons@7.25.1
+
 ## 7.25.0
 
 ### Patch Changes